refactor/refactor_wallet_base_backend

+++ hummingbot/wallet/wallet_base.pyx
@@ -6,22 +6,22 @@ cdef class WalletBase(NetworkIterator):
     def address(self) -> str:
         raise NotImplementedError
 
-    def get_balance(self, asset_name: str) -> float:
+    def get_balance(self, asset_name: str) -> Decimal:
         return self.c_get_balance(asset_name)
 
     def get_raw_balance(self, asset_name: str) -> int:
         return self.c_get_raw_balance(asset_name)
 
-    def get_all_balances(self) -> Dict[str, float]:
+    def get_all_balances(self) -> Dict[str, Decimal]:
         raise NotImplementedError
 
-    def send(self, address: str, asset_name: str, amount: float) -> str:
+    def send(self, address: str, asset_name: str, amount: Decimal) -> str:
         return self.c_send(address, asset_name, amount)
 
-    def to_nominal(self, asset_name: str, raw_amount: int) -> float:
+    def to_nominal(self, asset_name: str, raw_amount: int) -> Decimal:
         raise NotImplementedError
 
-    def to_raw(self, asset_name: str, nominal_amount: float) -> int:
+    def to_raw(self, asset_name: str, nominal_amount: Decimal) -> int:
         raise NotImplementedError
 
     cdef double c_get_balance(self, str asset_name) except? -1:
